requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
sqlalchemy==2.0.34
pydantic==2.8.2
httpx==0.27.2
apscheduler==3.10.4
jinja2==3.1.4
pika==1.3.2
circuitbreaker==2.0.0
strawberry-graphql==0.235.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
respx==0.21.1
pytest-asyncio==0.23.7
httpx==0.27.2
python-json-logger==2.0.7
prometheus-fastapi-instrumentator==7.0.0
google-generativeai==0.8.3
httpx-sse==0.4.0
websockets>=10.0
# v2.3.2 AI Update
